fix: rename folder on database repository
