+++ src/infrastructure/infrastructure.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { CreateUrlRepositoryService } from './database/repositories/url-repository/create-url-repository.service';
+import { CreateUrlRepositoryService } from './database/repositories/url/create-url-repository.service';
 import { UrlModel } from './database/models/url.model';
 import { DatabaseErrorHandler } from './database/utils/db-error-handler';
